added SmartUserAuth logic to endpoints

+++ backend/SmartSocial/src/main/java/de/hsbi/smartsocial/Controller/GroupController.java
@@ -1,5 +1,6 @@
 package de.hsbi.smartsocial.Controller;
 
+import de.fhbielefeld.smartuser.annotations.SmartUserAuth;
 import de.hsbi.smartsocial.Exceptions.GroupForMemberNotFoundException;
 import de.hsbi.smartsocial.Exceptions.GroupJoinException;
 import de.hsbi.smartsocial.Exceptions.GroupNotFoundException;
@@ -12,11 +13,15 @@ import de.hsbi.smartsocial.Service.GroupService;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.*;
+import jakarta.ws.rs.container.ContainerRequestContext;
+import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 
 import java.time.LocalDate;
 import java.util.List;
+
+import static de.hsbi.smartsocial.Service.UtilityService.isUserValid;
 
 /**
  * Author: Clemens Maas
@@ -49,6 +54,7 @@ public class GroupController {
         return Response.ok(group).build();
     }
 
+    @SmartUserAuth
     @GET
     @Path("/getById/{id}")
     @Produces(MediaType.APPLICATION_JSON)
@@ -61,6 +67,7 @@ public class GroupController {
         return Response.ok(group).build();
     }
 
+    @SmartUserAuth
     @GET
     @Path("/getByName/{name}")
     @Produces(MediaType.APPLICATION_JSON)
@@ -73,6 +80,7 @@ public class GroupController {
         return Response.ok(group).build();
     }
 
+    @SmartUserAuth
     @GET
     @Path("/getByUserId/{id}")
     @Produces(MediaType.APPLICATION_JSON)
@@ -85,6 +93,7 @@ public class GroupController {
         return Response.ok(groupmember).build();
     }
 
+    @SmartUserAuth
     @GET
     @Path("/getMembersByGroupId/{id}")
     @Produces(MediaType.APPLICATION_JSON)
@@ -97,6 +106,7 @@ public class GroupController {
         return Response.ok(groupmembers).build();
     }
 
+    @SmartUserAuth
     @GET
     @Path("/getMembersByGroupId_SV/{id}")
     @Produces(MediaType.APPLICATION_JSON)
@@ -109,6 +119,7 @@ public class GroupController {
         return Response.ok(groupmembers).build();
     }
 
+    @SmartUserAuth
     @GET
     @Path("/getAll")
     @Produces(MediaType.APPLICATION_JSON)
@@ -118,6 +129,7 @@ public class GroupController {
         return Response.ok(groups).build();
     }
 
+    @SmartUserAuth
     @GET
     @Path("/join/{userId}/{code}")
     @Produces(MediaType.APPLICATION_JSON)
@@ -130,6 +142,7 @@ public class GroupController {
         return Response.ok(group).build();
     }
 
+    @SmartUserAuth
     @POST
     @Path("/create")
     @Consumes(MediaType.APPLICATION_JSON)
@@ -143,6 +156,7 @@ public class GroupController {
         return Response.status(Response.Status.CREATED).entity(createdGroup).build();
     }
 
+    @SmartUserAuth
     @POST
     @Path("/update")
     @Consumes(MediaType.APPLICATION_JSON)
@@ -156,12 +170,20 @@ public class GroupController {
         return Response.ok(updatedGroup).build();
     }
 
+    @SmartUserAuth
     @DELETE
     @Path("/delete/{id}")
     @ApiResponse(responseCode = "204", description = "Deletes group")
-    public Response deleteGroup(@PathParam("id") Long id) {
-        groupService.deleteGroup(id);
-        return Response.status(Response.Status.NO_CONTENT).build();
+    public Response deleteGroup(@PathParam("id") Long id, @Context ContainerRequestContext requestContext) {
+        if (isUserValid(id, requestContext)) {
+//            Group group = groupService.findGroupById(id);
+//            if (requestContext.getSecurityContext().getUserPrincipal() == group.getAdminUser().getId() {
+//
+//            }
+            groupService.deleteGroup(id);
+            return Response.status(Response.Status.NO_CONTENT).build();
+        } else
+            return Response.status(Response.Status.FORBIDDEN).build();
     }
